Refactor INTRO out of dos_programs
